Atualização da compra de ações.

+++ app/src/main/java/br/com/stv/appbolsa/dao/IBuySellStock.kt
@@ -1,6 +1,7 @@
 package br.com.stv.appbolsa.dao
 
 import br.com.stv.appbolsa.model.Stock
+import java.math.BigDecimal
 import java.util.*
 
 interface IBuySellStock {
@@ -8,11 +9,11 @@ interface IBuySellStock {
     var note: String
     var stock: String
     var amount: Int
-    var cust: Double
-    var rates: Double
-    var averagePerStock: Double
-    var custOperation: Double
-    var totalAveragePerStock: Double
+    var cust: BigDecimal
+    var rates: BigDecimal
+    var averagePerStock: BigDecimal
+    var custOperation: BigDecimal
+    var totalAveragePerStock: BigDecimal
     var totalAmountStock: Int
     var updateDate: Date
 }
